+++ lw6/2_Patrushev_Oleg/ShellSort/MergeSorter.h
@@ -5,12 +5,12 @@ public:
 	MergeSorter(const std::vector<int>& vec);
 	~MergeSorter() = default;
 
-	virtual void MergeSort(int p, int r) = 0;
+	virtual void MergeSort(size_t p, size_t r) = 0;
 	std::vector<int> GetVector();
 
 public:
 
-	void Merge(int low, int mid, int high);
+	void Merge(size_t low, size_t mid, size_t high);
 
 	std::vector<int> m_vector;
 	size_t m_startVectorLength;
